+++ src/starknetCli.ts
@@ -8,7 +8,7 @@ import {
   IDeployAccountProps,
   IOptionalDebugInfo,
   IDeclareOptions,
-} from './index';
+} from './programFactory';
 import { encodeInputs } from './passes';
 import { CLIError, logCLIError } from './utils/errors';
 import { getDependencyGraph, hashFilename, reducePath } from './utils/postCairoWrite';
